003 Code/backend/capstone/requirement.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.54.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
gpg==1.10.0
h11==0.14.0
idna==3.4
inflection==0.5.1
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
uWSGI==2.0.22
wsproto==1.2.0
